+++ api/opds2/__init__.py
@@ -6,3 +6,4 @@ from .link import Link
 from .metadata import Metadata
 from .navigation import Navigation
 from .publication import Publication
+from .rights import Rights