+++ api/opds2/feed.py
@@ -3,6 +3,7 @@ from .metadata import Metadata
 from .navigation import Navigation
 from .publication import Publication
 from .link import Link
+from .rights import Rights
 
 
 class Feed:
@@ -11,7 +12,8 @@ class Feed:
         'navigation': Navigation,
         'links': Link,
         'groups': Group,
-        'publications': Publication
+        'publications': Publication,
+        'rights': Rights
     }
 
     def __init__(self, *args, **kwargs):
@@ -22,6 +24,7 @@ class Feed:
         self.groups = kwargs.get('groups', [])
         self.images = kwargs.get('images', [])
         self.facets = kwargs.get('facets', [])
+        self.rights = kwargs.get('rights', [])
 
     def addMetadata(self, metadata):
         self.metadata = self.componentizeObject('metadata', metadata)
@@ -39,6 +42,13 @@ class Feed:
     def addLinks(self, links):
         for link in links:
             self.addLink(link)
+
+    def addRight(self, right):
+        self.addComponent('rights', right)
+
+    def addRights(self, rights):
+        for right in rights:
+            self.addRight(right)
 
     def addPublication(self, publication):
         self.addComponent('publications', publication)
@@ -89,7 +99,7 @@ class Feed:
     def __dir__(self):
         return [
             'metadata', 'navigation', 'links', 'publications', 'groups',
-            'images', 'facets'
+            'images', 'rights', 'facets'
         ]
 
     def __iter__(self):