+++ tests/unit/test_opds_feed.py
@@ -27,6 +27,7 @@ class TestOPDSFacet:
         assert testFeed.groups == []
         assert testFeed.images == []
         assert testFeed.facets == []
+        assert testFeed.rights == []
 
     def test_addMetadata(self, testFeed, mocker):
         mockComponentize = mocker.patch.object(Feed, 'componentizeObject')
@@ -143,7 +144,7 @@ class TestOPDSFacet:
 
     def test_dir(self, testFeed):
         assert dir(testFeed) == [
-            'facets', 'groups', 'images', 'links', 'metadata', 'navigation', 'publications',
+            'facets', 'groups', 'images', 'links', 'metadata', 'navigation', 'publications', 'rights'
         ]
 
     def test_iter_success(self, testFeed, testIterableClass, mocker):