+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased version -- v0.11.3
 ### Added
 - New endpoint to update whole collection JSON objects in the API response
+- Publication rights data object to opds2 directory and collectionFetch endpoint
 ### Fixed
 - Add new DB entities for automatic collections
 - Add support for 'Most Recent' automatic collection